class5/homework/requirement.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
rank-bm25==0.2.2
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
jupyter==1.0.0
